Allow selection of alternative commit sets
The default commit set looks at the discussion items in the PR, and
selects the commit _before_ the last reviewer comment as the 'basis
commit'. This is the start point for the range of commits to be diffed,
in the same way that the target branch (usually master) is for the PR as
a whole.

Allow selecting any commit except the first or the last as this basis
commit. Once the commit is selected, the GitHub API call will be made as
usual and the diff updated.